feat: extract core logic into library crate for improved modularity
